requirements.txt
backports.zoneinfo==0.2.1
kafka-python==2.0.2
more-itertools==8.8.0
psycopg2-binary==2.9.1
SQLAlchemy==1.3.24
typing-extensions==3.10.0.2
